+++ lib/internal/modules/cjs/loader.js
@@ -216,6 +216,13 @@ function tryExtensions(p, exts, isMain) {
   return false;
 }
 
+function readExtensions() {
+  const exts = Object.keys(Module._extensions);
+  for (var i = 0; i < exts.length; ++i)
+    exts[i] = path.extname(exts[i]) || exts[i];
+  return exts;
+}
+
 var warned = false;
 Module._findPath = function(request, paths, isMain) {
   if (path.isAbsolute(request)) {
@@ -272,7 +279,7 @@ Module._findPath = function(request, paths, isMain) {
       if (!filename) {
         // try it with each of the extensions
         if (exts === undefined)
-          exts = Object.keys(Module._extensions);
+          exts = readExtensions();
         filename = tryExtensions(basePath, exts, isMain);
       }
     }
@@ -280,7 +287,7 @@ Module._findPath = function(request, paths, isMain) {
     if (!filename && rc === 1) {  // Directory.
       // try it with each of the extensions at "index"
       if (exts === undefined)
-        exts = Object.keys(Module._extensions);
+        exts = readExtensions();
       filename = tryPackage(basePath, exts, isMain);
       if (!filename) {
         filename = tryExtensions(path.resolve(basePath, 'index'), exts, isMain);
